made relevant changes
