+++ hangman/Python/Hangman/.gitignore
@@ -1,1 +1,2 @@
 .idea/
+Hangman1.py
